:rocket: Change database provider (PostgreSQL to MySQL).

+++ prisma/seed.ts
@@ -1,29 +1,31 @@
 import { hash } from "bcryptjs";
-import { PrismaClientKnownRequestError } from "@/lib/generated/prisma/internal/prismaNamespace";
 import { Prisma } from "@/lib/prisma";
-import { PrismaClientValidationError } from "@prisma/client/runtime/client";
 
+// prettier-ignore
 (async () => {
   try {
-    console.log(`🌱 Seeding database...`);
+    console.log("🌱 Seeding database...");
+
+    if (!process.env.SUPERUSER_EMAIL || !process.env.SUPERUSER_PASSWORD) {
+      throw new Error("Nama pengguna dan kata sandi admin harus disetel dalam variabel lingkungan.");
+    }
+
     await Prisma.$connect();
     await Prisma.pengguna.deleteMany();
-    await Prisma.pengguna.createMany({
-      data: [
-        {
-          nama_lengkap: "Administrator",
-          surel: process.env.SUPERUSER_EMAIL as string,
-          kata_sandi: await hash(process.env.SUPERUSER_PASSWORD as string, 10),
-        },
-      ],
+    await Prisma.pengguna.upsert({
+      where: { surel: process.env.SUPERUSER_EMAIL },
+      update: {},
+      create: {
+        nama_lengkap: "Administrator",
+        surel: process.env.SUPERUSER_EMAIL as string,
+        kata_sandi: await hash(process.env.SUPERUSER_PASSWORD as string, 10),
+      },
     });
   } catch (error: unknown) {
-    if (error instanceof PrismaClientValidationError) console.error(`❌ [PrismaClientValidationError] Error seeding database: ${error.message}.`);
-    else if (error instanceof PrismaClientKnownRequestError) console.error(`❌ [PrismaClientKnownRequestError] Error seeding database: ${error.cause}.`);
-    else console.error(`❌ [Unknown Error] Error seeding database: ${error instanceof Error ? error.message : String(error)}.`);
+    console.error(`❌ Error seeding: ${error}`);
     process.exit(1);
   } finally {
-    console.log(`✅ Database seeded.`);
+    console.log("✅ Database seeded.");
     await Prisma.$disconnect();
   }
 })();
